backend/requirements.txt
fastapi
uvicorn
requests
python-dotenv
aiohttp
tenacity
pydantic
